# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Campus Marketplace API
  version: 1.0.0
  description: Baseline contract for hardened API routes.
servers:
  - url: http://localhost:5000
paths:
...
  /api/products/{id}:
    get:
      summary: Get product by id
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '200':
          description: Product detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      summary: Update mutable product fields (non-status)
      description: Status mutation is handled by `/api/products/{id}/status`.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProductId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProductRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '403':
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete a product
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteProductResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ProductId:
      in: path
      name: id
      required: true
      schema:
        type: string
  schemas:
    Product:
      type: object
      properties:
        _id:
          type: string
        title:
          type: string
        description:
          type: string
        price:
          type: number
        quantity:
          type: integer
        category:
          type: string
        condition:
          type: string
        status:
          type: string
        seller:
          type: object
      required: [_id, title, price, category, condition, seller]
    ProductResponse:
      type: object
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          $ref: '#/components/schemas/Product'
      required: [success, data]
...
    UpdateProductRequest:
      type: object
      description: Product field update request. Status is not allowed here.
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        quantity:
          type: integer
        category:
          type: string
        condition:
          type: string
        images:
          type: array
          items:
            type: string
            format: uri
...
    DeleteProductResponse:
      type: object
      properties:
        success:
          type: boolean
          enum: [true]
        message:
          type: string
      required: [success]
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          enum: [false]
        message:
          type: string
        code:
          type: string
        details:
          oneOf:
            - type: array
            - type: object